Ransomware-detection-system/backend/requirements.txt
watchdog==3.0.0
scikit-learn==1.0.2
yara-python==4.2.3
flask==2.0.3
flask-cors==3.0.10
psutil==5.9.0
numpy==1.22.0
pandas==1.4.0